+++ Pandora/Paradigm/Structure/Modification.hs
@@ -1,0 +1,4 @@
+module Pandora.Paradigm.Structure.Modification (module Exports) where
+
+import Pandora.Paradigm.Structure.Modification.Prefixed as Exports
+import Pandora.Paradigm.Structure.Modification.Comprehension as Exports