+++ CHANGELOG.md
@@ -353,3 +353,4 @@
 * Define `rotate` method of `Morphable` typeclass
 * Remove `T_` joint scheme in favor of `T_U`
 * Define `Prefixed` wrapper for data structures whose values can be prefixed
+* Move `Comprehension` and `Prefixed` modules from `Ability` to `Modification`